frontend/requirements.txt
streamlit
Pillow==8.3.2


